backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
psycopg2-binary
asyncpg
python-dotenv
passlib[bcrypt]
email-validator
python-multipart
python-jose[cryptography]